fix refactor controllers
