+++ app/application.rb
@@ -1,17 +1,13 @@
 require 'sinatra'
-require 'sinatra/flash'
 require 'omniauth'
 require 'omniauth-google-oauth2'
-require_relative '../db/database'
-require_relative './requesters/policies_by_email_requester'
-require_relative './requesters/create_policy_requester'
+require_relative './helpers/session_helper'
+require_relative './controllers/auth_controller'
+require_relative './controllers/policies_controller'
+require_relative './controllers/home_controller'
 
 class Application < Sinatra::Base
-  EXPIRATE_AFTER = 1200 # seconds
-  LOGIN_PATHS = %r{/(login|logout)}
-  AUTH_CALLBACK_PATHS = %r{/(auth/[^/]+/(callback|failure))}
-
-  register Sinatra::Flash
+  EXPIRES_SESSION = 1200 # seconds
 
   configure do
     enable :sessions
@@ -24,100 +20,12 @@ class Application < Sinatra::Base
     use Rack::Session::Cookie,
       key: 'my_app_session',
       secret: ENV['COOKIE_SECRET'],
-      expire_after: EXPIRATE_AFTER
+      expire_after: EXPIRES_SESSION
   end
 
-  use OmniAuth::Builder do
-    provider :google_oauth2, ENV['GOOGLE_KEY'], ENV['GOOGLE_SECRET'], scope: 'email'
-  end
+  helpers SessionHelper
 
-  before do
-    return if request.path =~ LOGIN_PATHS
-    return if request.path =~ AUTH_CALLBACK_PATHS
-    return if user_signed_in?
-
-    redirect '/login'
-  end
-
-  get '/' do
-    policies = PolicyByEmailRequester.execute(omniauth_auth_email)
-
-    if policies
-      return erb :'../views/home', layout: :application, locals: {
-        policies: policies,
-        email: omniauth_auth_email
-      }
-    end
-
-    erb :'../views/generic_error', layout: :application, locals: {
-      email: omniauth_auth_email
-    }
-  end
-
-  get '/login' do
-    erb :'../views/login', layout: :application,
-    locals: {
-      google_key: ENV['GOOGLE_KEY'],
-      csrf_token: request.env['rack.session']['csrf']
-    }
-  end
-
-  post '/logout' do
-    Session.expire_all(omniauth_auth_email)
-    redirect '/login'
-  end
-
-  get '/new_policy' do
-    erb :'../views/new_policy', layout: :application,
-      locals: {
-        email: omniauth_auth_email
-      }
-  end
-
-  post '/create_policy' do
-    response = CreatePolicyRequester.execute(params.merge(email: omniauth_auth_email))
-
-    if response && !response[:errors]
-      flash[:success] = 'Policy successfully created!'
-    else
-      flash[:failed] = 'Policy not created! Try again later'
-    end
-
-    redirect '/'
-  end
-
-  get '/auth/:provider/callback' do
-    content_type 'text/plain'
-
-    session[:omniauth_auth] = request.env['omniauth.auth'].to_hash
-    user_email = session[:omniauth_auth]['info']['email']
-
-    Session.expire_all(user_email)
-
-    Session.create(
-      session_id: session[:session_id],
-      email: user_email,
-      expires_at: EXPIRATE_AFTER.seconds.from_now
-    )
-
-    redirect '/'
-  end
-
-  get '/auth/:provider/failure' do
-    content_type 'text/plain'
-    'Failure -> callback error'
-  end
-
-  def user_signed_in?
-    if omniauth_auth_email
-      current_session = Session.active_session(omniauth_auth_email)
-      return true if current_session && current_session.session_id_correct?(session[:session_id])
-    end
-
-    false
-  end
-
-  def omniauth_auth_email
-    session.dig(:omniauth_auth, 'info', 'email')
-  end
+  use HomeController
+  use AuthController
+  use PoliciesController
 end
