+++ app/helpers/session_helper.rb
@@ -1,0 +1,14 @@
+module SessionHelper
+  def user_signed_in?
+    if omniauth_auth_email
+      current_session = Session.active_session(omniauth_auth_email)
+      return true if current_session && current_session.session_id_correct?(session[:session_id])
+    end
+
+    false
+  end
+
+  def omniauth_auth_email
+    session.dig(:omniauth_auth, 'info', 'email')
+  end
+end
